My work up to today.
I'm wanted to get all of the images and text on the screen and now I'm going to work on how to put them where I need them on the screen.

I'm used two copies to be able to use what's correct on one in order to dress up how my code looks.